feat(User.ts): add functionality to on method and test it
In spite of the fact that the course insturctor does not initalized the events property with default value I have to do that because my version crashed every time (this.events is undefined) when the on function was called. In the future I should rethink this mod from time to time to keep the original track.